refactor(search): split lexical search into cohesive siblings
Keep the public search facade and ranking orchestration in the directory module while separating candidate retrieval, scoring, history, query helpers, and tests by responsibility.

Preserve candidate SQL, score weights, deterministic tie-breaking, graded-history gating, and lexical-only behavior.

Fixes #918
