+++ crates/rag-rat-core/src/search/lexical/query.rs
@@ -1,0 +1,27 @@
+pub(super) fn fts_query(query: &str) -> String {
+    let terms = query_terms(query)
+        .into_iter()
+        .map(|term| format!("\"{}\"", term.replace('"', "\"\"")))
+        .collect::<Vec<_>>();
+    if terms.is_empty() { "\"\"".to_string() } else { terms.join(" OR ") }
+}
+
+pub(super) fn query_terms(query: &str) -> Vec<String> {
+    query
+        .split(|c: char| !c.is_alphanumeric() && c != '_' && c != '-')
+        .filter(|term| !term.is_empty())
+        .map(str::to_ascii_lowercase)
+        .collect()
+}
+
+pub(super) fn snippet(text: &str, query: &str) -> String {
+    let terms = query_terms(query);
+    let lines = text.lines().collect::<Vec<_>>();
+    let hit = lines.iter().position(|line| {
+        let lower = line.to_ascii_lowercase();
+        terms.iter().any(|term| lower.contains(term))
+    });
+    let start = hit.unwrap_or(0).saturating_sub(1);
+    let end = (start + 3).min(lines.len());
+    lines[start..end].join("\n")
+}
